requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Login>=0.6.3,<0.7
Flask-SQLAlchemy>=3.1.1,<4.0
SQLAlchemy>=2.0.30,<3.0
cryptography>=42.0.0
python-dotenv>=1.0.1
psycopg2-binary==2.9.9
WTForms==3.0.1
Flask-WTF==1.2.1
bleach==6.0.0
gunicorn>=21.2.0